requirements.txt
APScheduler==3.8.0
asyncio==3.4.3
autopep8==1.6.0
dnspython==2.1.0
eventkit==0.8.9
ib-insync==0.9.69
nest-asyncio==1.5.1
numpy==1.21.2
pycodestyle==2.8.0
pymongo==3.12.1
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
toml==0.10.2
tzlocal==2.1
